use 8 bits for numframes, instead of 32

+++ r_defs.h
@@ -381,7 +381,7 @@ typedef struct
 
 typedef struct
 {
-  int32_t numframes;
+  int8_t numframes;
   spriteframe_t *spriteframes;
 } spritedef_t;
